feat: DRY principle - Unify NotesManager
