+++ app/src/main/java/com/sjaindl/notesdemoapp/NotesViewModel.kt
@@ -5,46 +5,36 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.sjaindl.notesdemoapp.model.Note
-import com.sjaindl.notesdemoapp.model.ShareType
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
 class NotesViewModel(
-    private val shareableNotesManager: ShareableNotesManager,
-    private val unshareableNotesManager: UnshareableNotesManager,
+    private val notesManager: NotesManager,
 ): ViewModel() {
     private val _notes = MutableStateFlow<List<Note>>(emptyList())
     val notes = _notes.asStateFlow()
 
     fun loadNotes() = viewModelScope.launch {
-        _notes.value = shareableNotesManager.load() + unshareableNotesManager.load()
+        _notes.value = notesManager.load()
     }
 
     fun addNote(note: Note) {
         viewModelScope.launch {
-            if (note.shareType == ShareType.Shareable) {
-                shareableNotesManager.save(note = note)
-            } else {
-                unshareableNotesManager.save(note = note)
-            }
+            notesManager.save(note = note)
         }
         _notes.value += note
     }
 
     fun deleteNote(note: Note) {
         viewModelScope.launch {
-            if (note.shareType == ShareType.Shareable) {
-                shareableNotesManager.delete(note = note)
-            } else {
-                unshareableNotesManager.delete(note = note)
-            }
+            notesManager.delete(note = note)
         }
         _notes.value -= note
     }
 
     fun share(note: Note) {
-        shareableNotesManager.share(note = note)
+        notesManager.share(note = note)
     }
 
     class NotesViewModelFactory(
@@ -53,10 +43,7 @@ class NotesViewModel(
         ViewModelProvider.NewInstanceFactory() {
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
             return NotesViewModel(
-                shareableNotesManager = ShareableNotesManager(
-                    context = context,
-                ),
-                unshareableNotesManager = UnshareableNotesManager(
+                notesManager = NotesManager(
                     context = context,
                 ),
             ) as T
